move service dir into src/app
Having code outside the src/app directory tree does not seem to conform with Angular conventions. In particular, Angular CLI refuses to create files outside of src/app (tried it with `ng g service ‘../<ServiceName>’`). See also https://angular.io/guide/styleguide#overall-structural-guidelines
