+++ src/app/editor-tabs/ace.component.spec.ts
@@ -2,14 +2,14 @@ import { AceComponent } from './ace.component';
 import { tick, fakeAsync, async, ComponentFixture, TestBed, flush } from '@angular/core/testing';
 import { mock, when, instance, anyString, spy, verify } from 'ts-mockito';
 import { MessagingModule, MessagingService } from '@testeditor/messaging-service';
-import { DocumentService } from '../../service/document/document.service';
-import { SyntaxHighlightingService } from 'service/syntaxHighlighting/syntax.highlighting.service';
+import { DocumentService } from '../service/document/document.service';
+import { SyntaxHighlightingService } from '../service/syntaxHighlighting/syntax.highlighting.service';
 import { ViewChild, Component, getDebugNode, DebugElement } from '@angular/core';
-import { AceClientsideSyntaxHighlightingService } from 'service/syntaxHighlighting/ace.clientside.syntax.highlighting.service';
+import { AceClientsideSyntaxHighlightingService } from '../service/syntaxHighlighting/ace.clientside.syntax.highlighting.service';
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/observable/of';
 import 'rxjs/add/observable/throw';
-import { Conflict } from 'service/document/conflict';
+import { Conflict } from '../service/document/conflict';
 import { By } from '@angular/platform-browser';
 import { ModalModule } from 'ngx-bootstrap/modal';
 import { BrowserDynamicTestingModule } from '@angular/platform-browser-dynamic/testing';
