+++ src/app/editor-tabs/editor-tabs.module.ts
@@ -5,10 +5,10 @@ import { TabsModule, TooltipModule } from 'ngx-bootstrap';
 
 import { AceComponent } from './ace.component';
 import { EditorTabsComponent } from './editor-tabs.component';
-import { DocumentService } from '../../service/document/document.service';
-import { DocumentServiceConfig } from '../../service/document/document.service.config';
-import { SyntaxHighlightingService } from '../../service/syntaxHighlighting/syntax.highlighting.service';
-import { AceClientsideSyntaxHighlightingService } from '../../service/syntaxHighlighting/ace.clientside.syntax.highlighting.service';
+import { DocumentService } from '../service/document/document.service';
+import { DocumentServiceConfig } from '../service/document/document.service.config';
+import { SyntaxHighlightingService } from '../service/syntaxHighlighting/syntax.highlighting.service';
+import { AceClientsideSyntaxHighlightingService } from '../service/syntaxHighlighting/ace.clientside.syntax.highlighting.service';
 
 @NgModule({
   imports: [
