+++ src/app/service/index/xtext.index.service.config.ts
