+++ src/app/service/syntaxHighlighting/syntax.highlighting.service.ts
